+++ python/paddle/distributed/passes/pass_utils.py
@@ -1043,43 +1043,6 @@ def _pir_get_backward_op_type(all_ops, op_idx):
     return ["backward_w"]
 
 
-def _add_dependency(recorder_op, waiter_op, name):
-    '''
-    Add the extra event dependency of the two operators.
-    This function mainly aims for the cross-programs in pipeline parallelism,
-    especial for the 'send_v2' 'recv_v2' etc.
-    '''
-    if not recorder_op.has_attr("force_record_event"):
-        recorder_op.set_bool_attr("force_record_event", True)
-    recorder_op.set_str_attr("event_to_record", name)
-    waiter_op.set_str_array_attr("events_to_wait", [name])
-
-
-def _add_dependency_if_necessary(
-    type_to_ops, cur_job_type, next_job_type, op_idx, rst_idx, var_name
-):
-    if not (
-        ("backward" in cur_job_type and "send_backward" in next_job_type)
-        or ("recv_forward" in cur_job_type and "forward" in next_job_type)
-    ):
-        return
-
-    first_used_idx = None
-    first_used_op = None
-    for used_op in (
-        type_to_ops[next_job_type][op_idx].result(rst_idx).all_used_ops()
-    ):
-        used_idx = type_to_ops[next_job_type].index(used_op)
-        if first_used_idx is None or used_idx < first_used_idx:
-            first_used_idx = used_idx
-            first_used_op = used_op
-
-    if first_used_op is not None:
-        _add_dependency(
-            type_to_ops[cur_job_type][op_idx], first_used_op, var_name
-        )
-
-
 def _create_program_and_ops(program, job_type, chunk_id=None):
     if chunk_id is not None:
         program_name = f"{job_type}{chunk_id}"
@@ -1090,225 +1053,6 @@ def _create_program_and_ops(program, job_type, chunk_id=None):
     ops = cloned_program.global_block().ops
 
     return program_name, cloned_program, ops
-
-
-def _split_program_for_vpp(
-    program, num_model_chunks, oprole_names, split_bw=False
-):
-    place = _get_device()
-    if isinstance(place, paddle.framework.CUDAPlace):
-        place = paddle.framework.CUDAPlace(
-            paddle.distributed.ParallelEnv().dev_id
-        )
-    cur_place = paddle.base.libpaddle.Place()
-    cur_place.set_place(place)
-
-    def get_var_name(op_idx, result_idx):
-        result_value = all_ops[op_idx].result(result_idx)
-        all_used_ops = result_value.all_used_ops()
-        shadow_output_op_used = None
-        for op in all_used_ops:
-            if op.name() == "builtin.shadow_output":
-                shadow_output_op_used = op
-
-        if shadow_output_op_used is not None:
-            var_name = shadow_output_op_used.attrs()["output_name"]
-        else:
-            var_name = f"var_{op_idx}_{all_ops[op_idx].name()}_{result_idx}"
-        return var_name
-
-    def add_persistable_var(op_idx, program_type):
-        all_program_types = list(type_to_program.keys())
-        following_program_types = all_program_types[
-            all_program_types.index(program_type) + 1 :
-        ]
-        op_num_results = type_to_ops[program_type][op_idx].num_results()
-        op_name = type_to_ops[program_type][op_idx].name()
-
-        for idx in range(op_num_results):
-            var_name = None
-            for type in reversed(following_program_types):
-                op_result = type_to_ops[type][op_idx].result(idx)
-                if op_result.use_empty():
-                    continue
-
-                # if this op's output is used, create the persistable
-                # var to be used in other programs.
-                if var_name is None:
-                    if op_name in ["pd_op.data", "builtin.parameter"]:
-                        var_name = op_result.name
-                    else:
-                        var_name = get_var_name(op_idx, idx)
-                        if "var_" in var_name:
-                            paddle.pir.set_insertion_point_after(
-                                type_to_ops[program_type][op_idx]
-                            )
-                            paddle._C_ops.set_persistable_value(
-                                type_to_ops[program_type][op_idx].result(idx),
-                                var_name,
-                            )
-
-                _add_dependency_if_necessary(
-                    type_to_ops, program_type, type, op_idx, idx, var_name
-                )
-
-                program_block = type_to_program[type].global_block()
-                new_result_var = program_block.add_kwarg(
-                    var_name, op_result.type()
-                )
-                new_result_var.place_attr = cur_place
-                new_result_var.persistable = op_result.persistable
-                type_to_ops[type][op_idx].result(idx).replace_all_uses_with(
-                    new_result_var
-                )
-
-        for type in following_program_types:
-            type_to_ops[type][op_idx].erase()
-
-    type_to_program = OrderedDict()
-    type_to_ops = OrderedDict()
-
-    # Step1: create programs and ops for each type
-    if not split_bw:
-        chunk_ids = list(range(num_model_chunks))
-
-        # Forward process
-        for chunk_id in chunk_ids:
-            for job_type in ["recv_forward", "forward"]:
-                name, prog, ops = _create_program_and_ops(
-                    program, job_type, chunk_id
-                )
-                type_to_program[name] = prog
-                type_to_ops[name] = ops
-
-        # Backward process
-        for chunk_id in reversed(chunk_ids):
-            for job_type in ["backward", "send_backward"]:
-                name, prog, ops = _create_program_and_ops(
-                    program, job_type, chunk_id
-                )
-                type_to_program[name] = prog
-                type_to_ops[name] = ops
-
-        # Optimizer
-        name, prog, ops = _create_program_and_ops(program, "optimizer")
-        type_to_program[name] = prog
-        type_to_ops[name] = ops
-    else:
-        for type in oprole_names:
-            if type == "optimizer":
-                type_to_program["optimizer"] = program.clone()
-                type_to_ops["optimizer"] = (
-                    type_to_program["optimizer"].global_block().ops
-                )
-            else:
-                chunk_ids = list(range(num_model_chunks))
-                if "backward" in type:
-                    chunk_ids.reverse()
-                for chunk_id in chunk_ids:
-                    type_to_program[type + str(chunk_id)] = program.clone()
-                    type_to_ops[type + str(chunk_id)] = (
-                        type_to_program[type + str(chunk_id)].global_block().ops
-                    )
-
-    # Step2: delete the ops not belong to the type
-    # 1. delete ops
-    # 2. add persistable var used between multiple programs
-    all_ops = program.global_block().ops
-    chunk_ids = list(range(num_model_chunks))
-    bwd_pattern_ops_type = []
-
-    for idx in range(len(all_ops) - 1, -1, -1):
-        op = all_ops[idx]
-        op_role = op.op_role
-        op_chunk_id = op.chunk_id
-        # Step2.1: infer chunk_id for ops that don't have chunk_id
-        if op_role != int(OpRole.Optimize) and op_chunk_id == -1:
-            op_chunk_id = infer_chunk_id(idx, all_ops, False)
-            if op_chunk_id == -1:
-                raise ValueError(
-                    f"Cannot infer chunk_id for op {op.name()} at index {idx}"
-                )
-
-        # Step2.2: identify the job_type of the op
-        if op_role == int(OpRole.Optimize):
-            job_type = "optimizer"
-        elif op_role == int(OpRole.Backward) and split_bw:
-            if len(bwd_pattern_ops_type) == 0:
-                bwd_pattern_ops_type = _pir_get_backward_op_type(all_ops, idx)
-            job_type = bwd_pattern_ops_type.pop()
-        elif op_role == int(OpRole.Backward) and (not split_bw):
-            if op.name() == "pd_op.send_v2":
-                job_type = "send_backward"
-            else:
-                job_type = "backward"
-        elif op_role == int(OpRole.Forward):
-            if op.name() == "pd_op.recv_v2" and (not split_bw):
-                job_type = "recv_forward"
-            else:
-                job_type = "forward"
-        else:
-            raise ValueError(
-                f"The op[{op.name()}]'s op role: {op_role} isn't one of recv_forward, forward, backward, send_backward or Optimizer."
-            )
-
-        # Step2.3: delete ops not belong to the type
-        if not split_bw:
-            current_type = (
-                job_type
-                if job_type == "optimizer"
-                else job_type + str(op_chunk_id)
-            )
-
-            # Get the position of the current type in type_to_program
-            all_types = list(type_to_ops.keys())
-            current_idx = all_types.index(current_type)
-
-            # Delete all ops before the current type
-            for type_name in all_types[:current_idx]:
-                type_to_ops[type_name][idx].erase()
-        else:
-            for type in oprole_names:
-                if type == job_type:
-                    break
-                if type != "optimizer":
-                    for chunk_id in chunk_ids:
-                        type_to_ops[type + str(chunk_id)][idx].erase()
-                else:
-                    type_to_ops[type][idx].erase()
-
-            chunk_order = range(0, op_chunk_id)
-            if "backward" in job_type:
-                chunk_order = range(num_model_chunks - 1, op_chunk_id, -1)
-            for chunk_id in chunk_order:
-                type_to_ops[job_type + str(chunk_id)][idx].erase()
-
-        # Step2.4: add persistable var used between multiple programs
-        if job_type != "optimizer":
-            add_persistable_var(idx, job_type + str(op_chunk_id))
-
-    return list(type_to_program.keys()), list(type_to_program.values())
-
-
-def _pir_program_for_vpp(
-    program, num_model_chunks, split_bw=False, enable_send_recv_overlap=False
-):
-    _pir_overlap_send_recv(program)
-
-    oprole_names = [
-        "recv_forward",
-        "forward",
-        "backward",
-        "send_backward",
-        "optimizer",
-    ]
-    if split_bw:
-        oprole_names = ["forward", "backward_b", "backward_w", "optimizer"]
-
-    program_types, programs = _split_program_for_vpp(
-        program, num_model_chunks, oprole_names, split_bw=split_bw
-    )
-    return program_types, programs
 
 
 def _program_for_vpp(
