+++ python/paddle/distributed/passes/pipeline_scheduler_pass/pipeline_vpp.py
@@ -13,7 +13,9 @@
 # limitations under the License.
 
 import logging
+from collections import OrderedDict
 
+import paddle
 import paddle.distributed as dist
 from paddle.base import core
 from paddle.distributed.auto_parallel.static.operators.common import (
@@ -21,22 +23,21 @@ from paddle.distributed.auto_parallel.static.operators.common import (
     is_data_parallel_scale_op,
 )
 
+from ...auto_parallel.static.utils import OpRole
 from ...utils.log_utils import get_logger
 from ..pass_base import register_pass
 from ..pass_utils import (
-    _pir_program_for_vpp,
+    _create_program_and_ops,
+    _get_device,
+    _pir_get_backward_op_type,
+    _pir_overlap_send_recv,
     _pir_split_matmul_grad_to_matmul,
     _program_for_vpp,
     _program_for_vpp_split_bwk,
+    infer_chunk_id,
     split_matmul_grad_to_matmul,
 )
 from .pipeline_pass_base import PipelinePassBase
-
-RECV_FORWARD = "recv_forward"
-SEND_BACKWARD = "send_backward"
-FORWARD = "forward"
-BACKWARD = "backward"
-OPT = "optimizer"
 
 logger = get_logger(logging.INFO)
 
@@ -50,10 +51,10 @@ class PipelineVirtualPipelinePass(PipelinePassBase):
         self._forward_micro_step_counter = {}
         self._backward_micro_step_counter = {}
         self.jobs_in_stable_phase_in_pir = [
-            BACKWARD,
-            RECV_FORWARD,
-            SEND_BACKWARD,
-            FORWARD,
+            self.BACKWARD,
+            self.RECV_FORWARD,
+            self.SEND_BACKWARD,
+            self.FORWARD,
         ]
 
     def _record_fwd_micro_step(self, virtual_pp_rank):
@@ -110,7 +111,7 @@ class PipelineVirtualPipelinePass(PipelinePassBase):
         for micro_step in range(warmup_steps):
             virtual_pp_rank = _get_virtual_pp_rank(micro_step, forward=True)
             micro_batch_id = self._record_fwd_micro_step(virtual_pp_rank)
-            fw_job = core.Job(FORWARD + str(virtual_pp_rank))
+            fw_job = core.Job(self.FORWARD + str(virtual_pp_rank))
             fw_job.set_micro_batch_id(micro_batch_id)
             job_list.append(fw_job)
 
@@ -122,7 +123,7 @@ class PipelineVirtualPipelinePass(PipelinePassBase):
             fwd_micro_batch_id = self._record_fwd_micro_step(
                 fwd_virtual_pp_rank
             )
-            fwd_job = core.Job(FORWARD + str(fwd_virtual_pp_rank))
+            fwd_job = core.Job(self.FORWARD + str(fwd_virtual_pp_rank))
             fwd_job.set_micro_batch_id(fwd_micro_batch_id)
             job_list.append(fwd_job)
 
@@ -134,9 +135,11 @@ class PipelineVirtualPipelinePass(PipelinePassBase):
                 bwd_virtual_pp_rank
             )
             if real_split_backward:
-                bwd_job = core.Job(BACKWARD + "_b" + str(bwd_virtual_pp_rank))
+                bwd_job = core.Job(
+                    self.BACKWARD + "_b" + str(bwd_virtual_pp_rank)
+                )
             else:
-                bwd_job = core.Job(BACKWARD + str(bwd_virtual_pp_rank))
+                bwd_job = core.Job(self.BACKWARD + str(bwd_virtual_pp_rank))
             bwd_job.set_micro_batch_id(bwd_micro_batch_id)
             job_list.append(bwd_job)
 
@@ -144,9 +147,9 @@ class PipelineVirtualPipelinePass(PipelinePassBase):
             virtual_pp_rank = _get_virtual_pp_rank(micro_step, forward=False)
             micro_batch_id = self._record_bwd_micro_step(virtual_pp_rank)
             if real_split_backward:
-                bwd_job = core.Job(BACKWARD + "_b" + str(virtual_pp_rank))
+                bwd_job = core.Job(self.BACKWARD + "_b" + str(virtual_pp_rank))
             else:
-                bwd_job = core.Job(BACKWARD + str(virtual_pp_rank))
+                bwd_job = core.Job(self.BACKWARD + str(virtual_pp_rank))
             bwd_job.set_micro_batch_id(micro_batch_id)
             job_list.append(bwd_job)
             # TODO(lizhiyu): Inserting 'backward_b' and 'backward_w' interleavedly can decrease the memory,
@@ -166,18 +169,18 @@ class PipelineVirtualPipelinePass(PipelinePassBase):
                         and micro_batch_id == accumulate_steps - 1
                     ):
                         w_job = core.Job(
-                            BACKWARD
+                            self.BACKWARD
                             + "_w"
                             + str(chunk_id)
                             + self.reduce_comm_suffix
                         )
                     else:
-                        w_job = core.Job(BACKWARD + "_w" + str(chunk_id))
+                        w_job = core.Job(self.BACKWARD + "_w" + str(chunk_id))
                     w_job.set_micro_batch_id(micro_batch_id)
                     job_list.append(w_job)
         job_types = [job.type() for job in job_list]
         logger.debug(f"The VPP job list: {job_types}")
-        opt_job = core.Job(OPT)
+        opt_job = core.Job(self.OPT)
         job_list.append(opt_job)
         return job_list
 
@@ -225,10 +228,12 @@ class PipelineVirtualPipelinePass(PipelinePassBase):
             virtual_pp_rank = _get_virtual_pp_rank(micro_step, forward=True)
             micro_batch_id = self._record_fwd_micro_step(virtual_pp_rank)
             if not real_split_backward:
-                recv_fwd_job = core.Job(RECV_FORWARD + str(virtual_pp_rank))
+                recv_fwd_job = core.Job(
+                    self.RECV_FORWARD + str(virtual_pp_rank)
+                )
                 recv_fwd_job.set_micro_batch_id(micro_batch_id)
                 job_list.append(recv_fwd_job)
-            fw_job = core.Job(FORWARD + str(virtual_pp_rank))
+            fw_job = core.Job(self.FORWARD + str(virtual_pp_rank))
             fw_job.set_micro_batch_id(micro_batch_id)
             job_list.append(fw_job)
 
@@ -241,7 +246,7 @@ class PipelineVirtualPipelinePass(PipelinePassBase):
                 fwd_micro_batch_id = self._record_fwd_micro_step(
                     fwd_virtual_pp_rank
                 )
-                fwd_job = core.Job(FORWARD + str(fwd_virtual_pp_rank))
+                fwd_job = core.Job(self.FORWARD + str(fwd_virtual_pp_rank))
                 fwd_job.set_micro_batch_id(fwd_micro_batch_id)
                 job_list.append(fwd_job)
 
@@ -252,7 +257,9 @@ class PipelineVirtualPipelinePass(PipelinePassBase):
                 bwd_micro_batch_id = self._record_bwd_micro_step(
                     bwd_virtual_pp_rank
                 )
-                bwd_job = core.Job(BACKWARD + "_b" + str(bwd_virtual_pp_rank))
+                bwd_job = core.Job(
+                    self.BACKWARD + "_b" + str(bwd_virtual_pp_rank)
+                )
                 bwd_job.set_micro_batch_id(bwd_micro_batch_id)
                 job_list.append(bwd_job)
         else:
@@ -272,8 +279,8 @@ class PipelineVirtualPipelinePass(PipelinePassBase):
                     bwd_virtual_pp_rank
                 )
                 for job_type in self.jobs_in_stable_phase_in_pir:
-                    if job_type.startswith(FORWARD) or job_type.startswith(
-                        RECV_FORWARD
+                    if job_type.startswith(self.FORWARD) or job_type.startswith(
+                        self.RECV_FORWARD
                     ):
                         job = core.Job(job_type + str(fwd_virtual_pp_rank))
                         job.set_micro_batch_id(fwd_micro_batch_id)
@@ -286,12 +293,14 @@ class PipelineVirtualPipelinePass(PipelinePassBase):
             virtual_pp_rank = _get_virtual_pp_rank(micro_step, forward=False)
             micro_batch_id = self._record_bwd_micro_step(virtual_pp_rank)
             if real_split_backward:
-                bwd_job = core.Job(BACKWARD + "_b" + str(virtual_pp_rank))
+                bwd_job = core.Job(self.BACKWARD + "_b" + str(virtual_pp_rank))
                 bwd_job.set_micro_batch_id(micro_batch_id)
                 job_list.append(bwd_job)
             else:
-                bwd_job = core.Job(BACKWARD + str(virtual_pp_rank))
-                send_bwd_job = core.Job(SEND_BACKWARD + str(virtual_pp_rank))
+                bwd_job = core.Job(self.BACKWARD + str(virtual_pp_rank))
+                send_bwd_job = core.Job(
+                    self.SEND_BACKWARD + str(virtual_pp_rank)
+                )
                 bwd_job.set_micro_batch_id(micro_batch_id)
                 send_bwd_job.set_micro_batch_id(micro_batch_id)
                 job_list.append(bwd_job)
@@ -313,18 +322,18 @@ class PipelineVirtualPipelinePass(PipelinePassBase):
                         and micro_batch_id == accumulate_steps - 1
                     ):
                         w_job = core.Job(
-                            BACKWARD
+                            self.BACKWARD
                             + "_w"
                             + str(chunk_id)
                             + self.reduce_comm_suffix
                         )
                     else:
-                        w_job = core.Job(BACKWARD + "_w" + str(chunk_id))
+                        w_job = core.Job(self.BACKWARD + "_w" + str(chunk_id))
                     w_job.set_micro_batch_id(micro_batch_id)
                     job_list.append(w_job)
         job_types = [job.type() for job in job_list]
         logger.debug(f"The VPP job list: {job_types}")
-        opt_job = core.Job(OPT)
+        opt_job = core.Job(self.OPT)
         job_list.append(opt_job)
         return job_list
 
@@ -510,7 +519,7 @@ class PipelineVirtualPipelinePass(PipelinePassBase):
         if split_backward:
             self._pir_split_matmul_grad_ops_to_matmul(program)
 
-        types, sub_program_list = _pir_program_for_vpp(
+        types, sub_program_list = self._pir_program_for_vpp(
             program, num_model_chunks, split_backward, enable_send_recv_overlap
         )
 
@@ -520,3 +529,230 @@ class PipelineVirtualPipelinePass(PipelinePassBase):
             )
 
         return types, sub_program_list
+
+    def _pir_program_for_vpp(
+        self,
+        program,
+        num_model_chunks,
+        split_bw=False,
+        enable_send_recv_overlap=False,
+    ):
+        _pir_overlap_send_recv(program)
+
+        oprole_names = [
+            "recv_forward",
+            "forward",
+            "backward",
+            "send_backward",
+            "optimizer",
+        ]
+        if split_bw:
+            oprole_names = ["forward", "backward_b", "backward_w", "optimizer"]
+
+        program_types, programs = self._split_program_for_vpp(
+            program, num_model_chunks, oprole_names, split_bw=split_bw
+        )
+        return program_types, programs
+
+    def _split_program_for_vpp(
+        self, program, num_model_chunks, oprole_names, split_bw=False
+    ):
+        place = _get_device()
+        if isinstance(place, paddle.framework.CUDAPlace):
+            place = paddle.framework.CUDAPlace(
+                paddle.distributed.ParallelEnv().dev_id
+            )
+        cur_place = paddle.base.libpaddle.Place()
+        cur_place.set_place(place)
+
+        def get_var_name(op_idx, result_idx):
+            result_value = all_ops[op_idx].result(result_idx)
+            all_used_ops = result_value.all_used_ops()
+            shadow_output_op_used = None
+            for op in all_used_ops:
+                if op.name() == "builtin.shadow_output":
+                    shadow_output_op_used = op
+
+            if shadow_output_op_used is not None:
+                var_name = shadow_output_op_used.attrs()["output_name"]
+            else:
+                var_name = f"var_{op_idx}_{all_ops[op_idx].name()}_{result_idx}"
+            return var_name
+
+        def add_persistable_var(op_idx, program_type):
+            all_program_types = list(type_to_program.keys())
+            following_program_types = all_program_types[
+                all_program_types.index(program_type) + 1 :
+            ]
+            op_num_results = type_to_ops[program_type][op_idx].num_results()
+            op_name = type_to_ops[program_type][op_idx].name()
+
+            for idx in range(op_num_results):
+                var_name = None
+                for type in reversed(following_program_types):
+                    op_result = type_to_ops[type][op_idx].result(idx)
+                    if op_result.use_empty():
+                        continue
+
+                    # if this op's output is used, create the persistable
+                    # var to be used in other programs.
+                    if var_name is None:
+                        if op_name in ["pd_op.data", "builtin.parameter"]:
+                            var_name = op_result.name
+                        else:
+                            var_name = get_var_name(op_idx, idx)
+                            if "var_" in var_name:
+                                paddle.pir.set_insertion_point_after(
+                                    type_to_ops[program_type][op_idx]
+                                )
+                                paddle._C_ops.set_persistable_value(
+                                    type_to_ops[program_type][op_idx].result(
+                                        idx
+                                    ),
+                                    var_name,
+                                )
+
+                    self._add_dependency_if_necessary(
+                        type_to_ops, program_type, type, op_idx, idx, var_name
+                    )
+
+                    program_block = type_to_program[type].global_block()
+                    new_result_var = program_block.add_kwarg(
+                        var_name, op_result.type()
+                    )
+                    new_result_var.place_attr = cur_place
+                    new_result_var.persistable = op_result.persistable
+                    type_to_ops[type][op_idx].result(idx).replace_all_uses_with(
+                        new_result_var
+                    )
+
+            for type in following_program_types:
+                type_to_ops[type][op_idx].erase()
+
+        type_to_program = OrderedDict()
+        type_to_ops = OrderedDict()
+
+        # Step1: create programs and ops for each type
+        if not split_bw:
+            chunk_ids = list(range(num_model_chunks))
+
+            # Forward process
+            for chunk_id in chunk_ids:
+                for job_type in ["recv_forward", "forward"]:
+                    name, prog, ops = _create_program_and_ops(
+                        program, job_type, chunk_id
+                    )
+                    type_to_program[name] = prog
+                    type_to_ops[name] = ops
+
+            # Backward process
+            for chunk_id in reversed(chunk_ids):
+                for job_type in ["backward", "send_backward"]:
+                    name, prog, ops = _create_program_and_ops(
+                        program, job_type, chunk_id
+                    )
+                    type_to_program[name] = prog
+                    type_to_ops[name] = ops
+
+            # Optimizer
+            name, prog, ops = _create_program_and_ops(program, "optimizer")
+            type_to_program[name] = prog
+            type_to_ops[name] = ops
+        else:
+            for type in oprole_names:
+                if type == "optimizer":
+                    type_to_program["optimizer"] = program.clone()
+                    type_to_ops["optimizer"] = (
+                        type_to_program["optimizer"].global_block().ops
+                    )
+                else:
+                    chunk_ids = list(range(num_model_chunks))
+                    if "backward" in type:
+                        chunk_ids.reverse()
+                    for chunk_id in chunk_ids:
+                        type_to_program[type + str(chunk_id)] = program.clone()
+                        type_to_ops[type + str(chunk_id)] = (
+                            type_to_program[type + str(chunk_id)]
+                            .global_block()
+                            .ops
+                        )
+
+        # Step2: delete the ops not belong to the type
+        # 1. delete ops
+        # 2. add persistable var used between multiple programs
+        all_ops = program.global_block().ops
+        chunk_ids = list(range(num_model_chunks))
+        bwd_pattern_ops_type = []
+
+        for idx in range(len(all_ops) - 1, -1, -1):
+            op = all_ops[idx]
+            op_role = op.op_role
+            op_chunk_id = op.chunk_id
+            # Step2.1: infer chunk_id for ops that don't have chunk_id
+            if op_role != int(OpRole.Optimize) and op_chunk_id == -1:
+                op_chunk_id = infer_chunk_id(idx, all_ops, False)
+                if op_chunk_id == -1:
+                    raise ValueError(
+                        f"Cannot infer chunk_id for op {op.name()} at index {idx}"
+                    )
+
+            # Step2.2: identify the job_type of the op
+            if op_role == int(OpRole.Optimize):
+                job_type = "optimizer"
+            elif op_role == int(OpRole.Backward) and split_bw:
+                if len(bwd_pattern_ops_type) == 0:
+                    bwd_pattern_ops_type = _pir_get_backward_op_type(
+                        all_ops, idx
+                    )
+                job_type = bwd_pattern_ops_type.pop()
+            elif op_role == int(OpRole.Backward) and (not split_bw):
+                if op.name() == "pd_op.send_v2":
+                    job_type = "send_backward"
+                else:
+                    job_type = "backward"
+            elif op_role == int(OpRole.Forward):
+                if op.name() == "pd_op.recv_v2" and (not split_bw):
+                    job_type = "recv_forward"
+                else:
+                    job_type = "forward"
+            else:
+                raise ValueError(
+                    f"The op[{op.name()}]'s op role: {op_role} isn't one of recv_forward, forward, backward, send_backward or Optimizer."
+                )
+
+            # Step2.3: delete ops not belong to the type
+            if not split_bw:
+                current_type = (
+                    job_type
+                    if job_type == "optimizer"
+                    else job_type + str(op_chunk_id)
+                )
+
+                # Get the position of the current type in type_to_program
+                all_types = list(type_to_ops.keys())
+                current_idx = all_types.index(current_type)
+
+                # Delete all ops before the current type
+                for type_name in all_types[:current_idx]:
+                    type_to_ops[type_name][idx].erase()
+            else:
+                for type in oprole_names:
+                    if type == job_type:
+                        break
+                    if type != "optimizer":
+                        for chunk_id in chunk_ids:
+                            type_to_ops[type + str(chunk_id)][idx].erase()
+                    else:
+                        type_to_ops[type][idx].erase()
+
+                chunk_order = range(0, op_chunk_id)
+                if "backward" in job_type:
+                    chunk_order = range(num_model_chunks - 1, op_chunk_id, -1)
+                for chunk_id in chunk_order:
+                    type_to_ops[job_type + str(chunk_id)][idx].erase()
+
+            # Step2.4: add persistable var used between multiple programs
+            if job_type != "optimizer":
+                add_persistable_var(idx, job_type + str(op_chunk_id))
+
+        return list(type_to_program.keys()), list(type_to_program.values())
